Added controllers for listings
